+++ src/test/java/se/sundsvall/casedata/api/model/PatchErrandTest.java
@@ -1,5 +1,17 @@
 package se.sundsvall.casedata.api.model;
 
+import org.hamcrest.MatcherAssert;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import se.sundsvall.casedata.api.model.validation.enums.CaseType;
+import se.sundsvall.casedata.integration.db.model.enums.Priority;
+
+import java.time.LocalDate;
+import java.time.OffsetDateTime;
+import java.util.List;
+import java.util.Random;
+import java.util.UUID;
+
 import static com.google.code.beanmatchers.BeanMatchers.hasValidBeanConstructor;
 import static com.google.code.beanmatchers.BeanMatchers.hasValidBeanEquals;
 import static com.google.code.beanmatchers.BeanMatchers.hasValidBeanHashCode;
@@ -10,19 +22,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.hamcrest.CoreMatchers.allOf;
 import static se.sundsvall.casedata.TestUtil.createExtraParametersList;
 import static se.sundsvall.casedata.TestUtil.createFacility;
-
-import java.time.LocalDate;
-import java.time.OffsetDateTime;
-import java.util.List;
-import java.util.Random;
-import java.util.UUID;
-
-import org.hamcrest.MatcherAssert;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-
-import se.sundsvall.casedata.api.model.validation.enums.CaseType;
-import se.sundsvall.casedata.integration.db.model.enums.Priority;
 
 class PatchErrandTest {
 
@@ -58,6 +57,7 @@ class PatchErrandTest {
 		final var extraParameters = createExtraParametersList();
 		final var facilities = List.of(createFacility());
 		final var suspension = new Suspension();
+		final var relatesTo = List.of(new RelatedErrand());
 
 		// Act
 		final var result = PatchErrand.builder()
@@ -74,6 +74,7 @@ class PatchErrandTest {
 			.withExtraParameters(extraParameters)
 			.withFacilities(facilities)
 			.withSuspension(suspension)
+			.withRelatesTo(relatesTo)
 			.build();
 
 		// Assert
@@ -91,6 +92,7 @@ class PatchErrandTest {
 		assertThat(result.getExtraParameters()).isEqualTo(extraParameters);
 		assertThat(result.getFacilities()).isEqualTo(facilities);
 		assertThat(result.getSuspension()).isEqualTo(suspension);
+		assertThat(result.getRelatesTo()).isEqualTo(relatesTo);
 	}
 
 	@Test